Add: ArithmeticException to Int and Double

+++ SDKAndroid/src/main/java/app/cybrid/sdkandroid/core/AssetPipe.kt
@@ -11,7 +11,7 @@ class AssetPipe {
         }
 
         fun transform(value:String, asset: AssetBankModel, unit: String) : BigDecimal {
-            return transformAny(BigDecimal(value.toInt()), asset, unit)
+            return transformAny(BigDecimal(value), asset, unit)
         }
 
         fun transform(value:Int, asset: AssetBankModel, unit: String) : BigDecimal {
